+++ ethereum/permissioning/src/test/java/org/hyperledger/besu/ethereum/permissioning/account/AccountPermissioningControllerFactoryTest.java
@@ -16,14 +16,10 @@ package org.hyperledger.besu.ethereum.permissioning.account;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.AssertionsForClassTypes.fail;
-import static org.assertj.core.api.ThrowableAssert.catchThrowable;
-import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.Mockito.when;
 
 import org.hyperledger.besu.datatypes.Address;
 import org.hyperledger.besu.ethereum.permissioning.LocalPermissioningConfiguration;
 import org.hyperledger.besu.ethereum.permissioning.PermissioningConfiguration;
-import org.hyperledger.besu.ethereum.permissioning.SmartContractPermissioningConfiguration;
 import org.hyperledger.besu.ethereum.transaction.TransactionSimulator;
 import org.hyperledger.besu.metrics.noop.NoOpMetricsSystem;
 import org.hyperledger.besu.plugin.services.MetricsSystem;
@@ -62,7 +58,7 @@ public class AccountPermissioningControllerFactoryTest {
     assertThat(localConfig.isAccountAllowlistEnabled()).isFalse();
 
     PermissioningConfiguration permissioningConfiguration =
-        new PermissioningConfiguration(Optional.of(localConfig), Optional.empty());
+        new PermissioningConfiguration(Optional.of(localConfig));
 
     Optional<AccountPermissioningController> controller =
         AccountPermissioningControllerFactory.create(
@@ -80,7 +76,7 @@ public class AccountPermissioningControllerFactoryTest {
     assertThat(localConfig.isAccountAllowlistEnabled()).isTrue();
 
     PermissioningConfiguration permissioningConfiguration =
-        new PermissioningConfiguration(Optional.of(localConfig), Optional.empty());
+        new PermissioningConfiguration(Optional.of(localConfig));
 
     Optional<AccountPermissioningController> controller =
         AccountPermissioningControllerFactory.create(
@@ -91,93 +87,6 @@ public class AccountPermissioningControllerFactoryTest {
 
     Assertions.assertThat(controller).isNotEmpty();
     assertThat(controller.get().getAccountLocalConfigPermissioningController()).isNotEmpty();
-    assertThat(controller.get().getTransactionSmartContractPermissioningController()).isEmpty();
-  }
-
-  @Test
-  public void createOnchainConfigWithAccountPermissioningDisabledShouldReturnEmpty() {
-    SmartContractPermissioningConfiguration onchainConfig =
-        SmartContractPermissioningConfiguration.createDefault();
-    assertThat(onchainConfig.isSmartContractAccountAllowlistEnabled()).isFalse();
-
-    PermissioningConfiguration permissioningConfiguration =
-        new PermissioningConfiguration(Optional.empty(), Optional.of(onchainConfig));
-
-    Optional<AccountPermissioningController> controller =
-        AccountPermissioningControllerFactory.create(
-            permissioningConfiguration,
-            transactionSimulator,
-            metricsSystem,
-            Collections.emptyList());
-
-    Assertions.assertThat(controller).isEmpty();
-  }
-
-  @Test
-  public void createOnchainConfigOnlyControllerShouldReturnExpectedController() {
-    SmartContractPermissioningConfiguration onchainConfig = onchainConfig();
-    assertThat(onchainConfig.isSmartContractAccountAllowlistEnabled()).isTrue();
-
-    PermissioningConfiguration permissioningConfiguration =
-        new PermissioningConfiguration(Optional.empty(), Optional.of(onchainConfig));
-
-    Optional<AccountPermissioningController> controller =
-        AccountPermissioningControllerFactory.create(
-            permissioningConfiguration,
-            transactionSimulator,
-            metricsSystem,
-            Collections.emptyList());
-
-    Assertions.assertThat(controller).isNotEmpty();
-    assertThat(controller.get().getAccountLocalConfigPermissioningController()).isEmpty();
-    assertThat(controller.get().getTransactionSmartContractPermissioningController()).isNotEmpty();
-  }
-
-  @Test
-  public void createOnchainShouldFailIfValidationFails() {
-    SmartContractPermissioningConfiguration onchainConfig = onchainConfig();
-    assertThat(onchainConfig.isSmartContractAccountAllowlistEnabled()).isTrue();
-
-    PermissioningConfiguration permissioningConfiguration =
-        new PermissioningConfiguration(Optional.empty(), Optional.of(onchainConfig));
-
-    when(transactionSimulator.processAtHead(any())).thenThrow(new RuntimeException());
-
-    final Throwable thrown =
-        catchThrowable(
-            () ->
-                AccountPermissioningControllerFactory.create(
-                    permissioningConfiguration,
-                    transactionSimulator,
-                    metricsSystem,
-                    Collections.emptyList()));
-
-    assertThat(thrown)
-        .isInstanceOf(IllegalStateException.class)
-        .hasMessage("Error validating onchain account permissioning smart contract configuration");
-  }
-
-  @Test
-  public void createLocalAndOnchainControllerShouldReturnExpectedControllers() {
-    LocalPermissioningConfiguration localConfig = localConfig();
-    assertThat(localConfig.isAccountAllowlistEnabled()).isTrue();
-
-    SmartContractPermissioningConfiguration onchainConfig = onchainConfig();
-    assertThat(onchainConfig.isSmartContractAccountAllowlistEnabled()).isTrue();
-
-    PermissioningConfiguration permissioningConfiguration =
-        new PermissioningConfiguration(Optional.of(localConfig), Optional.of(onchainConfig));
-
-    Optional<AccountPermissioningController> controller =
-        AccountPermissioningControllerFactory.create(
-            permissioningConfiguration,
-            transactionSimulator,
-            metricsSystem,
-            Collections.emptyList());
-
-    Assertions.assertThat(controller).isNotEmpty();
-    assertThat(controller.get().getAccountLocalConfigPermissioningController()).isNotEmpty();
-    assertThat(controller.get().getTransactionSmartContractPermissioningController()).isNotEmpty();
   }
 
   private LocalPermissioningConfiguration localConfig() {
@@ -188,15 +97,6 @@ public class AccountPermissioningControllerFactoryTest {
     localPermissioningConfiguration.setAccountPermissioningConfigFilePath(
         createTempFile().getPath());
     return localPermissioningConfiguration;
-  }
-
-  private SmartContractPermissioningConfiguration onchainConfig() {
-    SmartContractPermissioningConfiguration onchainPermissioningConfiguration =
-        SmartContractPermissioningConfiguration.createDefault();
-    onchainPermissioningConfiguration.setAccountSmartContractAddress(
-        Address.fromHexString("0x0000000000000000000000000000000000008888"));
-    onchainPermissioningConfiguration.setSmartContractAccountAllowlistEnabled(true);
-    return onchainPermissioningConfiguration;
   }
 
   private File createTempFile() {